Feature #22588 : Komplette Kosten Nutzen Rechnung wird ausgeführt und Ergebnis angezeigt, Funktion wurde in Thread realisiert und automatisiert bei Task hinzufügen + löschen

+++ client/src/main/java/at/ac/tuwien/sepm/assignment/group02/client/rest/CostBenefitController.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface CostBenefitController {
 
-    int costValueFunction(List<TaskDTO> taskList) throws PersistenceLayerException;
+    double costValueFunction(List<TaskDTO> taskList) throws PersistenceLayerException;
 
 }
